CSCMETAX-61: [ADD] Modify update_common_info for bulk create and update
Before, fields modified_by_api and created_by_api (timestamps) were included in the fields
before serializer.is_valid(). If is_valid() raised an error, the timestamp would be included
in the object that is returned to the user, in which case the timestamps can seem confusing.
Especially in the case of bulk create, where the object is not actually created, but the
created_by_api timestamp is still present.

Now, timestamps are added to the data AFTER is_valid(), circumventing the previous issue.