CSCMETAX-61: [ADD] Redis: Always GET from from localhost first, before trying out other slaves or master
Rationale: When running the app clustered, the 'randomization' will have already happened
when the request has reached the python code, in which case it is beneficial to always
try reading from localhost first, and only in case of errors try other instances.

Write operations still need to go directly to master.