CSCMETAX-251: [FIX] Fall back to ORM query instead of sql when deleting directories
Previously when switching from ORM to sql was probably due to some other errors,
since now it seems to work OK again. This change is not motivated by wanting to use
ORM instead of sql though, but because there was a gargantuan problem that rose
with calling self.track_fields() and getattr(self, field_name) inside Common model
__init__, which only manifested itself when using Directory.objects.raw() to query
objects... Something to do with inheritance, getattr(), and infinite recursion.

The reason why Model.objects.raw() works differently is still unclear, but
using the ORM query works, and wouldn't have been a bad choice in the first place
anyway. In the future when/if needing to do sql queries, it might be better to do std
sql query using db.connection and cursor.execute(), get the ids, and then ORM
query using those received ids, which is going to be two queries of course instead
of just one, but at least will work...