CSCMETAX-442: [FIX|REF] Fix datacite convert errors. Convert metax datasets to datacite xml using 'datacite' library
Previous implemenetation of datacite conversion relied on xml transformation, which used our custom 'metax dataset
xml' as a starting point, and contained no xml schema validation at the end of it. The end result currently is quite
far from the schema, and the xslt is very difficult to get right.

As an alternative solution, use python 'datacite' library, which sports a datacite json-schema, and offers conversion
from that json to xml representation. The json-way is easier to validate (there is a all-in-one json schema available),
and easier to modify (just python code to create datacite json from metax dataset, instead of screwing around with
xslt files). Just to be sure, the produced xml was manually tested to validate agains the datacite xml.

The 'datacite' library conveniently also includes wrappers for DOI-related api's.