CSCMETAX-388: [ADD] Better way to execute code at the very end of a request.
During a request, enables any code, anywhere to append callable objects to a list
of callables that will get executed in order of adding, at the end of the request.
No need to deliver data or parameters from some code deepdown all the way to the
out-most region in the code, where the request is almost finished, since the
callables contain all the information the need, and are self-contained.

If needed tho, it would be very easy to pass additional parameters to the callable,
such as the request- or the response-object.