CSCMETAX-61: [FIX] When deleting datasets, only publish 'delete' to rabbitmq
Before the recent refactoring of publishing to rabbitmq, when deleting a dataset,
only 'delete' was published. Now, after the refactoring, it was discovered that
when deleting a dataset, what happens is that actually the CR.save() method is
executed as well, by the superclass. Previously, it did not cause a rabbitmq
publish due to it being done in a different place. This is now fixed by having
the superclass (model Common) only save the value of field 'removed', and calling
its superclass (not CR.save(), and not common.save())