TPASPKT-977 Cache Finto data in 'index_refdata'
'index_refdata' is performed periodically (eg. during CI workflow),
but it's slow and consumes one thread for several minutes parsing XML.
Cache the generated Graph instances using `pickle` to greatly reduce
runtime when the source reference data has not changed since the last
run.

Cache invalidation is performed by downloading the file and calculating
its checksum. If we have a corresponding cache file, we can load
the corresponding graph from cache and use that instead. If not, the
reference data is downloaded again and old cache files for the data type
are deleted.